refactor(score): engine.ts を teamBuilder と simulation に分割 (re-export 互換)
